requirements.txt
streamlit

python-telegram-bot==22.1
flask 
requests